create paginated users to wrap in suspense
